Massive reorganisation
Blogged here:

Efficient Emacs.org ➞ .el tangling
Holger's Computer Calisthenics and Orthodontia
http://www.holgerschurig.de/en/emacs-init-tangle/

and expanded here:

https://bitbucket.org/holgerschurig/emacsconf/src

The native org mode tangling is replaced by custom elisp code that is
blazingly fast. There is less code in the actual org file and the
order of tangled code is same as in the org file.